Made some modifications to the Approval so that it knows exactly what versions of every file are being sent for approval Added the following columns:   * date_created - so we know when the file was created   * renamed workflow_version to just "version", because everything has a version,  this is the version of the request.   * workflow_hash - this is just a quick way to see what files and versions are associated with the request, it could be factored out.   * study - a quick relationship link to the study, so that this model is easier to use.   * workflow - ditto   * approval_files - these is a list from a new link table that links an approval to specific files and versions.
The RequestApproval is logically sound, but still needs some additional pieces in place to be callable from a BPMN workflow diagram.

Altered the file service to pick up on changes to files vs adding new files, so that versions are picked up correctly as
users modify their submission - adding new files or replacing existing ones.  Deleting files worries me, and I will need to revisit this.

The damn base test keeps giving me a headache, so I made changes there to see if clearing and dropping the database each time won't allow the tests to pass more consistently.

Lots more tests around the file service to make sure it is versioning user uploaded files correctly.

The "Test Request Approval Script" tries to find to assure the correct behavior as this is likely to be called many times repeatedly and with little knowledge of the internal system.  So it should just "do the right thing".