The master workflow now generates a dictionary of dictionaries, rather than a dictionary of strings. Each entry in status is a dictionary with 'status' and 'message' keys. We updated the StudyService._update_status_of_workflow_meta method to accommodate the master workflow changes.
Changed the order of some if statements so we move forward while we have good data, and run all the error states at the end.
Added some comments to explain the cascading if statements

We changed the names of private methods to begin with one underscore, so they work in the test environment.
Changed  some internal calls to accommodate the underscore change.