Add some additional checks in the workflow service so we assure the data we are creating during validation can be serialized and deserialized just as it would be through the API.  (Was hanging us up on dates)
Assure that if we generate a default value for a date in the task data, it is stored as an ISO String.

remove any unserializable data from the task_data when an error is encountered, rather than just dropping all the task_data. This case seems to happen a lot and it leaves us with nothing to go on.